fix: make json-converters public
