+++ src/pkgs/Altinn.Register.Contracts/src/Altinn.Register.Contracts/LangCode.cs
@@ -5,7 +5,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Text;
-using System.Text.Json;
+using Altinn.Register.Contracts.JsonConverters;
 using Altinn.Swashbuckle.Filters;
 using CommunityToolkit.Diagnostics;
 
@@ -15,7 +15,7 @@ namespace Altinn.Register.Contracts;
 /// Represents a language code.
 /// </summary>
 [DebuggerDisplay("{Code}")]
-[JsonConverter(typeof(LangCode.JsonConverter))]
+[JsonConverter(typeof(LangCodeJsonConverter))]
 [SwaggerString]
 public sealed class LangCode
     : IEquatable<LangCode>
@@ -144,6 +144,11 @@ public sealed class LangCode
     /// </summary>
     public string Code => _code;
 
+    /// <summary>
+    /// Gets the language code as UTF-8 bytes.
+    /// </summary>
+    internal ReadOnlySpan<byte> Utf8 => _utf8.AsSpan();
+
     private LangCode(string code, ReadOnlySpan<byte> utf8)
     {
         Guard.IsNotNull(code);
@@ -220,89 +225,5 @@ public sealed class LangCode
     public static bool operator <=(LangCode left, LangCode right)
     {
         return left.CompareTo(right) <= 0;
-    }
-
-    /// <summary>
-    /// <see cref="System.Text.Json.Serialization.JsonConverter"/> for <see cref="LangCode"/>.
-    /// </summary>
-    internal sealed class JsonConverter
-        : JsonConverter<LangCode>
-    {
-        /// <inheritdoc cref="JsonConverter{T}.Read(ref Utf8JsonReader, Type, JsonSerializerOptions)"/>
-        public static LangCode? Read(ref Utf8JsonReader reader)
-        {
-            if (reader.TokenType == JsonTokenType.Null)
-            {
-                return null;
-            }
-
-            if (reader.TokenType is not JsonTokenType.String)
-            {
-                throw new JsonException($"Expected a string, but got {reader.TokenType}.");
-            }
-
-            if (reader.ValueIsEscaped || reader.HasValueSequence)
-            {
-                var str = reader.GetString();
-                if (str is null)
-                {
-                    return null;
-                }
-
-                return FromCode(str);
-            }
-
-            return FromCode(reader.ValueSpan);
-        }
-
-        /// <inheritdoc cref="JsonConverter{T}.ReadAsPropertyName(ref Utf8JsonReader, Type, JsonSerializerOptions)"/>
-        public static LangCode ReadAsPropertyName(ref Utf8JsonReader reader)
-        {
-            if (reader.TokenType is not JsonTokenType.PropertyName)
-            {
-                throw new JsonException($"Expected a string, but got {reader.TokenType}.");
-            }
-
-            if (reader.ValueIsEscaped || reader.HasValueSequence)
-            {
-                var str = reader.GetString();
-                if (str is null)
-                {
-                    throw new JsonException($"Expected a property name, but read null");
-                }
-
-                return FromCode(str);
-            }
-
-            return FromCode(reader.ValueSpan);
-        }
-
-        /// <inheritdoc cref="JsonConverter{T}.Write(Utf8JsonWriter, T, JsonSerializerOptions)"/>
-        public static void Write(Utf8JsonWriter writer, LangCode value)
-        {
-            writer.WriteStringValue(value._utf8.AsSpan());
-        }
-
-        /// <inheritdoc cref="JsonConverter{T}.WriteAsPropertyName(Utf8JsonWriter, T, JsonSerializerOptions)"/>
-        public static void WriteAsPropertyName(Utf8JsonWriter writer, LangCode value)
-        {
-            writer.WritePropertyName(value._utf8.AsSpan());
-        }
-
-        /// <inheritdoc/>
-        public override LangCode? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            => Read(ref reader);
-
-        /// <inheritdoc/>
-        public override LangCode ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            => ReadAsPropertyName(ref reader);
-
-        /// <inheritdoc/>
-        public override void Write(Utf8JsonWriter writer, LangCode value, JsonSerializerOptions options)
-            => Write(writer, value);
-
-        /// <inheritdoc/>
-        public override void WriteAsPropertyName(Utf8JsonWriter writer, [DisallowNull] LangCode value, JsonSerializerOptions options)
-            => WriteAsPropertyName(writer, value);
     }
 }
